Uniform buffer working, staged write fix
The staged writing wasn't destroying the staging buffer, now it does. To
do so it waits on the copy_queue which can be anoying, better to do it
manually when in need of performance, thanks to pipeline barriers.

Also descriptor sets, uniform buffers are finally tested and working !
Rotating triangle goes brrrrr